Split the pause-point wait status poll loop into its own file
pause_point_wait.go grew to 548 lines across PR-2's review-response
commits (94cb67c9, 9a5fc981), tripping
TestProductionGoFilesStayFocused's 500-line architecture gate and
turning PR #1926's CI red (build-cli, Go CLI Security Analysis, and
test-windows-installers). This was not caught locally because the
architecture test lives in the separate release-automation Go module
(cli/release-automation), which scripts/check-go-cli.sh does not run
from the project-runner working directory, and it was not re-run after
the review-response commits landed.

Move the trigger-arming and status-poll-loop functions
(waitForPausePoint, pausePointIsArmed, waitForPausePointStatus,
queryPausePointStatusAtTimeout, pausePointWaitStateForStatus, plus the
pausePointWaitState type and its supporting const/var) into a new
pause_point_wait_poll.go file. Pure file split: no behavior change, no
public signature change. pause_point_wait.go keeps command entry
points, option parsing, and response normalization/formatting.
